Applying band-aid for fallout from SWITCHYARD-1411
The getRoute() method on CamelComponentImplementationModel has been out-of-date
for some time now (schema hasn't supported it for months).  The commit for
SWITCHYARD-1411 in components removed the method, which created a compilation
error in tools.  My guess is it's not used anywhere, but I went ahead and
changed the affected to call getXMLPath() which returns the path to the XML
route definition in a SY app.